62 copying project on infra and deploy should be optional

+++ utils/infra.go
@@ -8,11 +8,12 @@ import (
 )
 
 type InfraChange struct {
-	OutputDir   string
-	ProfilePath string
-	CmdOnDir    func(string, string, string, bool)
-	Verbosity   bool
-	tmpDir      string
+	OutputDir    string
+	ProfilePath  string
+	CmdOnDir     func(string, string, string, bool)
+	Verbosity    bool
+	tmpDir       string
+	CopyToTmpDir bool
 }
 
 func (infraChange *InfraChange) DeployClientAndServer() {
@@ -21,28 +22,28 @@ func (infraChange *InfraChange) DeployClientAndServer() {
 	infraChange.CmdOnDir(
 		fmt.Sprintf("omgd run task set-ip-to-profile --profile=%s", infraChange.ProfilePath),
 		"",
-		fmt.Sprintf("%s/.omgdtmp", infraChange.OutputDir),
+		infraChange.OutputDir,
 		infraChange.Verbosity,
 	)
 
 	infraChange.CmdOnDir(
 		fmt.Sprintf("omgd build-templates --profile=%s", infraChange.ProfilePath),
 		"",
-		fmt.Sprintf("%s/.omgdtmp", infraChange.OutputDir),
+		infraChange.OutputDir,
 		infraChange.Verbosity,
 	)
 
 	infraChange.CmdOnDir(
 		fmt.Sprintf("omgd build-clients --profile=%s", infraChange.ProfilePath),
 		"",
-		fmt.Sprintf("%s/.omgdtmp", infraChange.OutputDir),
+		infraChange.OutputDir,
 		infraChange.Verbosity,
 	)
 
 	infraChange.CmdOnDir(
 		fmt.Sprintf("omgd run nakama-server --profile=%s", infraChange.ProfilePath),
 		"",
-		fmt.Sprintf("%s/.omgdtmp", infraChange.OutputDir),
+		infraChange.OutputDir,
 		infraChange.Verbosity,
 	)
 }
@@ -54,7 +55,7 @@ func (infraChange *InfraChange) DeployInfra() {
 	infraChange.CmdOnDir(
 		fmt.Sprintf("omgd run task deploy-infra --profile=%s", infraChange.ProfilePath),
 		"",
-		fmt.Sprintf("%s/.omgdtmp", infraChange.OutputDir),
+		infraChange.OutputDir,
 		infraChange.Verbosity,
 	)
 }
@@ -66,7 +67,7 @@ func (infraChange *InfraChange) DestroyInfra() {
 	infraChange.CmdOnDir(
 		fmt.Sprintf("omgd run task destroy-infra --profile=%s", infraChange.ProfilePath),
 		"",
-		fmt.Sprintf("%s/.omgdtmp", infraChange.OutputDir),
+		infraChange.OutputDir,
 		infraChange.Verbosity,
 	)
 }
@@ -75,47 +76,52 @@ func (infraChange *InfraChange) setup() {
 	infraChange.ProfilePath = strings.ReplaceAll(infraChange.ProfilePath, "profiles/", ".omgd/")
 
 	// 1. Should create or empty .omgdtmp directory to work in
-	infraChange.tmpDir = fmt.Sprintf("%s/.omgdtmp", infraChange.OutputDir)
+	if infraChange.CopyToTmpDir {
+		infraChange.tmpDir = infraChange.OutputDir
+		infraChange.tmpDir = fmt.Sprintf("%s/.omgdtmp", infraChange.OutputDir)
 
-	if infraChange.OutputDir == "." {
-		infraChange.tmpDir = ".omgdtmp"
-	}
+		if infraChange.OutputDir == "." {
+			infraChange.tmpDir = ".omgdtmp"
+		}
 
-	_, err := os.Stat(infraChange.tmpDir)
-	if !os.IsNotExist(err) {
-		err = os.RemoveAll(infraChange.tmpDir)
+		_, err := os.Stat(infraChange.tmpDir)
+		if !os.IsNotExist(err) {
+			err = os.RemoveAll(infraChange.tmpDir)
+
+			if err != nil {
+				log.Fatal(err)
+			}
+		}
+
+		err = os.Mkdir(infraChange.tmpDir, 0755)
 
 		if err != nil {
 			log.Fatal(err)
 		}
+
+		// 2. Should clone repo at base of dir (? how to test w/o submodules? clone entire base repo maybe?)
+		sccp := StaticCodeCopyPlan{
+			skipPaths: []string{
+				infraChange.tmpDir,
+				".git",
+			},
+		}
+
+		cwd, err := os.Getwd()
+
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		if infraChange.OutputDir == "." {
+			sccp.CopyStaticDirectory(cwd, infraChange.tmpDir)
+		} else {
+			sccp.CopyStaticDirectory(infraChange.OutputDir, infraChange.tmpDir)
+		}
+
+		infraChange.OutputDir = infraChange.tmpDir
 	}
 
-	err = os.Mkdir(infraChange.tmpDir, 0755)
-
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	// 2. Should clone repo at base of dir (? how to test w/o submodules? clone entire base repo maybe?)
-	sccp := StaticCodeCopyPlan{
-		skipPaths: []string{
-			infraChange.tmpDir,
-			".git",
-		},
-	}
-
-	cwd, err := os.Getwd()
-
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	if infraChange.OutputDir == "." {
-		sccp.CopyStaticDirectory(cwd, infraChange.tmpDir)
-	} else {
-		sccp.CopyStaticDirectory(infraChange.OutputDir, infraChange.tmpDir)
-	}
-
-	// 3. Build profiles directory in new .omgdtmp dir
-	BuildProfiles(infraChange.tmpDir, infraChange.Verbosity)
+	// 3. Build profiles directory
+	BuildProfiles(infraChange.OutputDir, infraChange.Verbosity)
 }
